+++ config/main.php
@@ -144,11 +144,6 @@ $config =
             'class' => '\skeeks\cms\components\CurrentSite',
         ],
 
-        'dbDump' =>
-        [
-            'class' => '\skeeks\cms\components\db\DbDumpComponent',
-        ],
-
         'console' =>
         [
             'class' => 'skeeks\cms\components\ConsoleComponent',